+++ smos-query/src/Smos/Query/OptParse.hs
@@ -164,7 +164,6 @@ combineToInstructions SmosQueryConfig {..} c Flags {..} Environment {..} mc =
                     fromMaybe defaultProjection $
                       combineMaybe (<>) (mwc workConfProjection) workFlagProjection,
                   workSetSorter = mwc workConfSorter <|> workFlagSorter,
-                  workSetSmartMode = fromMaybe False $ workFlagSmartMode <|> mwc workConfSmartMode,
                   workSetHideArchive = hideArchiveWithDefault HideArchive workFlagHideArchive
                 }
         CommandProjects ProjectsFlags {..} ->
@@ -292,12 +291,8 @@ parseCommandWork = info parser modifier
                 <*> parseFilterOptionsRel
                 <*> parseProjectionArgs
                 <*> parseSorterArgs
-                <*> parseSmartModeFlag
                 <*> parseHideArchiveFlag
             )
-
-parseSmartModeFlag :: Parser (Maybe Bool)
-parseSmartModeFlag = flag Nothing (Just True) (mconcat [long "smart", help "Smart mode"])
 
 parseCommandWaiting :: ParserInfo Command
 parseCommandWaiting = info parser modifier